% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnPCA-package.R
\docType{package}
\name{LearnPCA-package}
\alias{LearnPCA-package}
\alias{LearnPCA}
\title{Functions, Data Sets and Vignettes to Aid in Learning Principal Components Analysis (PCA)}
\description{
PASTE DESCRIPTION HERE
}
\author{
Bryan A. Hanson and David T. Harvey.

Maintainer: Bryan A. Hanson \email{hanson@depauw.edu}
}
\keyword{multivariate}
\keyword{package}
