% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCAtoXhat.R
\name{PCAtoXhat}
\alias{PCAtoXhat}
\title{Use PCA Results to Reconstruct All or Part of the Original Data Set}
\usage{
PCAtoXhat(pca, ncomp = NULL)
}
\arguments{
\item{pca}{An object of class \code{prcomp}.  The results of data reduction by PCA.}

\item{ncomp}{Integer.  The number of principal components to use in reconstructing
the data set.  Must be no larger than the number of variables.}
}
\value{
A matrix with the same dimensions as \code{pca$x} (the dimensions of the
original data set).
}
\description{
This function allows one to reconstruct an approximation (\code{Xhat}) of the
original data
using some or all of the principal components, starting from the results of PCA.
Inspired by and follows \url{https://stackoverflow.com/a/23603958/633251} very closely.
We are grateful for this post by StackOverflow contributor Marc in the box.
}
\examples{
# Example data from ?prcomp (see discussion at Stats.StackExchange.com/q/397793)
C <- chol(S <- toeplitz(.9 ^ (0:31)))
set.seed(17)
X <- matrix(rnorm(32000), 1000, 32)
Z <- X \%*\% C

pcaz <- prcomp(Z)
tst <- PCAtoXhat(pcaz)
mean(tst - Z)

# Plot to show the effect of increasing ncomp

ntests <- ncol(Z)
val1 <- rep(NA_real_, ntests)
val2 <- rep(NA_real_, ntests)
for (i in 1:ntests) {
	ans <- PCAtoXhat(pcaz, i)
	del <- ans - Z
	val1[i] <- sqrt(sum(del^2)/length(del)) # RMSD
	val2[i] <- mean(del)
}
plot(val1, type = "b", ylim = c(0.0, max(val1)),
  main = "Root Mean Squared Deviation", xlab = "No. of Components Retained", ylab = "RMSD")
abline(h = 0.0, col = "pink")
me <- .Machine$double.eps
plot(val2, type = "b", ylim = c(-me, me), main = "Residual Error",
  xlab = "No. of Components Retained", ylab = "Mean Error")
thres <- c(-me, 0.0, me)
abline(h = thres, col = "pink") # all values below .Machine$double.eps

}
