% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnPCA-package.R
\docType{package}
\name{LearnPCA-package}
\alias{LearnPCA-package}
\alias{LearnPCA}
\title{Functions, Data Sets and Vignettes to Aid in Learning Principal Components Analysis (PCA)}
\description{
Principal component analysis (PCA) is one of the most widely used data analysis techniques.
This package provides a series of vignettes explaining PCA starting from basic concepts.
The primary purpose is to serve as a self-study resource for anyone wishing to understand PCA better.
A few convenience functions are provided as well.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bryanhanson.github.io/LearnPCA/}
  \item Report bugs at \url{https://github.com/bryanhanson/LearnPCA/issues}
}

}
\author{
Bryan A. Hanson and David T. Harvey.

Maintainer: Bryan A. Hanson \email{hanson@depauw.edu}
}
\keyword{multivariate}
\keyword{package}
