% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCsearch.R
\name{PCsearch}
\alias{PCsearch}
\title{Shiny application to demonstrate the search for the 1st two principal components
for a randomly generated set of data.}
\usage{
PCsearch()
}
\description{
@return  None.  A web page opens with the application running.
@author Bryan A. Hanson, David T. Harvey
}
